Move tests out of published package
Fixes #51
