+++ Cargo.toml
@@ -9,12 +9,7 @@ homepage = "https://github.com/image-rs/image-gif"
 repository = "https://github.com/image-rs/image-gif"
 documentation = "https://docs.rs/gif"
 edition = "2021"
-exclude = [
-    "tests/crashtest/*",
-    "tests/samples/*",
-    "benches/*.gif",
-    "gif-afl/*",
-]
+include = ["src/**", "LICENSE-*", "README.md", "benches/*.rs"]
 
 [lib]
 bench = false
@@ -36,26 +31,6 @@ raii_no_panic = []
 color_quant = ["dep:color_quant"]
 # Reservation for a feature turning off std
 std = []
-
-[[test]]
-name = "check_testimages"
-required-features = ["std"]
-
-[[test]]
-name = "crashtest"
-required-features = ["std"]
-
-[[test]]
-name = "decode"
-required-features = ["std"]
-
-[[test]]
-name = "stall"
-required-features = ["std"]
-
-[[test]]
-name = "roundtrip"
-required-features = ["std"]
 
 [[bench]]
 name = "decode"
