+++ tests/decode.rs
@@ -1,4 +1,25 @@
+#![cfg(feature = "std")]
+
 use gif::{Decoder, DecodeOptions, DisposalMethod, Encoder, Frame};
+use std::fs::File;
+
+#[test]
+fn test_simple_indexed() {
+    let mut decoder = Decoder::new(File::open("tests/samples/sample_1.gif").unwrap()).unwrap();
+    let frame = decoder.read_next_frame().unwrap().unwrap();
+    assert_eq!(&*frame.buffer, &[
+        1, 1, 1, 1, 1, 2, 2, 2, 2, 2,
+        1, 1, 1, 1, 1, 2, 2, 2, 2, 2,
+        1, 1, 1, 1, 1, 2, 2, 2, 2, 2,
+        1, 1, 1, 0, 0, 0, 0, 2, 2, 2,
+        1, 1, 1, 0, 0, 0, 0, 2, 2, 2,
+        2, 2, 2, 0, 0, 0, 0, 1, 1, 1,
+        2, 2, 2, 0, 0, 0, 0, 1, 1, 1,
+        2, 2, 2, 2, 2, 1, 1, 1, 1, 1,
+        2, 2, 2, 2, 2, 1, 1, 1, 1, 1,
+        2, 2, 2, 2, 2, 1, 1, 1, 1, 1
+    ][..]);
+}
 
 #[test]
 fn frame_consistency_is_configurable() {
